feat: 标签作废增加校验，如果存在有效数据则禁止作废
git-svn-id: svn://172.16.27.209/gsgh/gsgh/lghjft-backend@513 63aaa323-bd20-4782-bdcf-f624a29d9816

+++ yudao-module-lghjft/src/main/java/cn/iocoder/yudao/module/lghjft/service/hj/bqgl/BqglServiceImpl.java
@@ -26,6 +26,9 @@ import java.time.LocalDate;
 
 import com.baomidou.mybatisplus.core.conditions.update.LambdaUpdateWrapper;
 
+import static cn.iocoder.yudao.framework.common.exception.util.ServiceExceptionUtil.exception;
+import static cn.iocoder.yudao.module.lghjft.enums.ErrorCodeConstants.BQ_HAS_VALID_DATA;
+
 /**
  * 标签管理 Service 实现类
  *
@@ -64,6 +67,13 @@ public class BqglServiceImpl implements BqglService {
 
     @Override
     public void deleteBqxx(String id) {
+        // 校验是否存在有效数据
+        Long count = ghHjBqMapper.selectCount(new LambdaQueryWrapper<GhHjBqDO>()
+                .eq(GhHjBqDO::getBqId, id)
+                .ge(GhHjBqDO::getYxqz, LocalDate.now()));
+        if (count > 0) {
+            throw exception(BQ_HAS_VALID_DATA);
+        }
         ghDmHjBqMapper.deleteById(id);
     }
